* Add changes and refactor
