+++ .gitignore
@@ -101,3 +101,4 @@ modules.xml
 # Sonarlint plugin
 .idea/sonarlint
 /.idea/
+/target/
