+++ src/main/java/org/apache/logging/log4j/kusto/MainTest.java
@@ -4,6 +4,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 public class MainTest {
+
     private static final Logger logger = LogManager.getLogger(MainTest.class);
 
     public static void main(String[] args) {
